short int, not long nit pick to match the rest
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@3888 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/VotingArbiter.cxx
@@ -289,7 +289,7 @@ bool VotingArbiter::isPollSureToFail(void) const
   return false;
 }
 
-unsigned long int VotingArbiter::timeRemaining(void) const
+unsigned short int VotingArbiter::timeRemaining(void) const
 {
   if (_votingBooth == NULL) {
     return 0;
@@ -304,7 +304,7 @@ unsigned long int VotingArbiter::timeRemaining(void) const
   if (remaining < 0.0f) {
     return 0;
   }
-  return (unsigned int)remaining;
+  return (unsigned short int)remaining;
 }
 
 bool VotingArbiter::retractVote(std::string player)
